update: with_llm also move to JacLLM
